src/backend/requirements.txt
# PostProber AI Backend Dependencies

# FastAPI - Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# LangChain - AI framework
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.20

# OpenAI
openai==1.10.0

# Environment variables
python-dotenv==1.0.1

# Validation
pydantic==2.5.3

# Async HTTP (for health monitoring)
aiohttp==3.9.1

# Background jobs (for future health monitoring)
apscheduler==3.10.4

# Redis (for caching, future use)
redis==5.0.1

# WebSocket support
websockets==12.0

# CORS
python-multipart==0.0.6
